fix bug, use correct types, add gio.ToInt64()
